requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
requests